chore: add a Nix package definition for `midnight-node`
This requires small changes in `midnight-ledger` to make it work with
`cargo-vendor`. The changes against `main` can be seen in
<https://github.com/midnightntwrk/midnight-ledger/pull/81> there.

Since `midnight-node` depends on 2 tags of the ledger repository:
* `ledger-6.1.0-alpha.4`, and
* `hard-fork-test-ledger-6.1.0-alpha.4`,

I have created two branches there (one for hard-fork, one normal):

* [fix/cargo-vendor--ledger-6.1.0-alpha.4](https://github.com/blockfrost/midnight-ledger/tree/fix/cargo-vendor--ledger-6.1.0-alpha.4), and
* [fix/cargo-vendor--hard-fork-test-ledger-6.1.0-alpha.4](https://github.com/blockfrost/midnight-ledger/tree/fix/cargo-vendor--hard-fork-test-ledger-6.1.0-alpha.4),

against which this repository can be built, just to show it’s possible:

```
❯ nix build -L
[…]

❯ ./result/bin/midnight-node --version
midnight-node 0.17.1
```

The next release, `ledger-6.1.0-alpha.5`, will probably contain this
ledger patch, if it gets merged to their `main` in the meantime.